+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * Add `buffer` function
 * Add `Signal#buffer` method
 * Add `apply` function
+* Add `Signal.throwError` static method
 
 ## 2.0.0 (2019-01-12)